+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2020-08-26
+
+### General
+- Hide pcap_on_demand behind feature flag.
+
+### Fix
+- Update is public now.
+
+## [0.3.3] - 2020-08-26
+
+### Fixed
+- EvalConfig, TimeRepresentation, TimeFormat are public now.
+
 ## [0.3.2] - 2020-04-27
 
 ### General